apps/api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6
fluids>=1.3
pint>=0.23
scipy>=1.12.0
numpy>=1.26.0
ruamel.yaml>=0.18

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dateutil>=2.8
